+++ tests/tests/Report/CoberturaTest.php
@@ -21,7 +21,7 @@ final class CoberturaTest extends TestCase
 
         $this->assertStringMatchesFormatFile(
             TEST_FILES_PATH . 'BankAccount-cobertura-line.xml',
-            $cobertura->process($this->getLineCoverageForBankAccount()->getReport())
+            $cobertura->process($this->getLineCoverageForBankAccount()->buildReport())
         );
     }
 
@@ -31,7 +31,7 @@ final class CoberturaTest extends TestCase
 
         $this->assertStringMatchesFormatFile(
             TEST_FILES_PATH . 'BankAccount-cobertura-path.xml',
-            $cobertura->process($this->getPathCoverageForBankAccount()->getReport())
+            $cobertura->process($this->getPathCoverageForBankAccount()->buildReport())
         );
     }
 
@@ -41,7 +41,7 @@ final class CoberturaTest extends TestCase
 
         $this->assertStringMatchesFormatFile(
             TEST_FILES_PATH . 'ignored-lines-cobertura.xml',
-            $cobertura->process($this->getCoverageForFileWithIgnoredLines()->getReport())
+            $cobertura->process($this->getCoverageForFileWithIgnoredLines()->buildReport())
         );
     }
 
@@ -51,7 +51,7 @@ final class CoberturaTest extends TestCase
 
         $this->assertStringMatchesFormatFile(
             TEST_FILES_PATH . 'class-with-anonymous-function-cobertura.xml',
-            $cobertura->process($this->getCoverageForClassWithAnonymousFunction()->getReport())
+            $cobertura->process($this->getCoverageForClassWithAnonymousFunction()->buildReport())
         );
     }
 
@@ -61,7 +61,7 @@ final class CoberturaTest extends TestCase
 
         $this->assertStringMatchesFormatFile(
             TEST_FILES_PATH . 'class-with-outside-function-cobertura.xml',
-            $cobertura->process($this->getCoverageForClassWithOutsideFunction()->getReport())
+            $cobertura->process($this->getCoverageForClassWithOutsideFunction()->buildReport())
         );
     }
 }